backend/order-processing/requirements.txt
mysql-connector-python==8.3.0
pika==1.3.2
